requirements.txt
aniso8601==8.0.0
autopep8==1.5.7
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
cryptography==35.0.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pytz==2020.1
requests==2.26.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.2
urllib3==1.26.7
Werkzeug==1.0.1
